Move utils/find_duplicates.py to ssg/_utils_duplicates.py
